Moved create and getParentAndFillAncestors methods from the page model to the page service
